+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added language color light & dark schemes [#1409](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1409)
 - Added language formatter [#1410](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1410)
 - Added default file template [#1411](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1411)
+- Added language folding support [#1412](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1412)
 
 ## [2025.1.4]